requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
defusedxml==0.7.1
Django==3.2.10
django-allauth==0.47.0
django-appconf==1.0.5
django-ckeditor==6.1.0
django-compressor==2.4.1
django-htmlmin==0.11.0
django-imagekit==4.1.0
django-jazzmin==2.4.8
django-js-asset==1.2.2
django-seo-js==0.3.3
django-taggit==1.5.1
django-widget-tweaks==1.4.8
django-yaturbo==1.0.0
html5lib==1.1
idna==3.3
oauthlib==3.1.1
pilkit==2.0
Pillow==8.4.0
pkg_resources==0.0.0
pycparser==2.21
PyJWT==2.3.0
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
rcssmin==1.0.6
requests==2.26.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.7
webencodings==0.5.1
